+++ src/mqt/bench/benchmarks/shor.py
@@ -27,7 +27,6 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:  # pragma: no cover
     from numpy.typing import NDArray
 import math
-from typing import cast
 
 import numpy as np
 from qiskit.circuit import Instruction, ParameterVector, QuantumCircuit, QuantumRegister
@@ -108,7 +107,7 @@ class Shor:
                 if bits_little_endian[j] == "1":
                     angles[i] += pow(2, -k)
 
-        return cast("NDArray[np.float64]", angles * np.pi)
+        return angles * np.pi
 
     @staticmethod
     def _phi_add_gate(angles: NDArray[np.float64] | ParameterVector) -> QuantumCircuit: